django-auth/requirements.txt
Django>=3.0,<4.0
djangorestframework_simplejwt>=4.4.0
djangorestframework>=3.11.1
django-cors-headers>=3.5.0
fasttext==0.9.2
